+++ derive_builder/src/options/struct_options.rs
@@ -21,9 +21,10 @@ pub struct StructOptions {
     /// Target struct name.
     pub build_target_ident: syn::Ident,
     /// Represents lifetimes and type parameters attached to the declaration of items.
-    ///
-    /// We assume that this is identical for the builder and its build target.
     pub generics: syn::Generics,
+    /// Represents lifetimes and type parameters attached to the `impl` 
+    /// block for the builder.
+    pub inherent_generics: syn::Generics,
     /// Emit deprecation notes to the user,
     /// e.g. if a deprecated attribute was used in `derive_builder`.
     pub deprecation_notes: DeprecationNotes,
@@ -46,6 +47,7 @@ impl StructOptions {
             ident: &self.builder_ident,
             derives: &self.derives,
             generics: Some(&self.generics),
+            impl_generics: Some(self.inherent_generics.split_for_impl().0),
             visibility: &self.builder_visibility,
             fields: Vec::with_capacity(self.struct_size_hint),
             functions: Vec::with_capacity(self.struct_size_hint),